rearrange functions; move FluxTimesXSHelper to be inner class of FluxDepletionOperator We move FluxTimesXSHelper to be an inner class of FluxDepletionOperator so we can avoid needing to copy the number and cross_sections attributes.
